requirements.txt
langchain==1.2.15
langchain_community==0.4.1
langchain_core==1.2.30
langchain_groq==1.1.2
langchain_huggingface==1.2.1
pydantic==2.13.1
python-dotenv==1.2.2
scikit_learn==1.8.0
streamlit
os
collections